⚗️✅ Divide typer test in other file
